refactor: streamline Chat component and introduce useStreamingChat hook
This update refactors the Chat component by removing unused state and logic, consolidating streaming functionality into a new custom hook, useStreamingChat. The new hook manages stream events, error handling, and user messages more efficiently, improving code readability and maintainability. Additionally, tests for the useStreamingChat hook have been added to ensure its functionality and reliability.
